Enable statistics for and during USB printing
The current code drops all USB from the statistics.
Moreover, the job statistics are not available during
USB priting. The support was not enabled for several
reasons - one of them being the missing handling of
print pauses silently executed by the USB Host.

Nevertheless, I prefer best effort counters instead
of none. I've therefore decided to extend the M73
command handler and trigger on 0% and 100% g-codes.
Additionally the statistics menu is whitelisted during
USB printing and shows the data of the current job.